refactor: move sign in logic to ProfileMenu component
